bump kotlin

+++ app/src/main/java/com/satis/app/feature/account/AccountState.kt
@@ -14,6 +14,6 @@ data class AccountState(
 
 @Parcelize
 data class BuildData(
-  val versionNum: Int,
+  val versionNum: Long,
   val buildTime: Long
 ) : Parcelable
